Split up configured IDPs away from the SP registration.
Per:
https://github.com/spring-projects/spring-security-saml/issues/436#issuecomment-505201556
"I know we have gone back and forth on this, but I'm afraid we may have
an issue at the moment for Saml2ServiceProviderRegistration having the
getIdentityProviders on it. The problem is that this will not scale if
there are very large number of IdPs. We need a lookup based on the alias
on a repository interface for the IdP as well."